requirements.txt
certifi==2017.11.5
chardet==3.0.4
idna==2.6
praw==5.3.0
prawcore==0.13.0
py==1.4.34
pyaml==17.12.1
pytest==3.1.2
PyYAML==3.12
requests==2.18.4
update-checker==0.16
urllib3==1.22
internetarchive
flask
apscheduler
python-dotenv
inflect
boto3